Using ZXing instead of IDAutomation for Barcode Generation
This was demo version of IDAutomation which has some missing fonts. As a result SAREE00003 was not getting printed correctly.

Problem description: This font looked correctly on my machine but when I looked on shop computer this font had some missing glyph, when we were printing then barcode was not getting printed.

As a workaround to this problem, I enerated label on my system, converted it to a doc file, printed this doc file to XPS and changed paper size to A4 during conversion, sent this XPS file to shop. This printed correctly. All other combination like generating PDF/A format and printing it (as this will contain embedded font)... was not working because of allignment mismatch.

Cause:
Cause of this problem was using demo version of IDAutomation font

Fix:
To fix this pronlem, we replaced IDAutomation workflow by generating barcode using zxing library. git location: https://github.com/micjahn/ZXing.Net, binary location: https://github.com/micjahn/ZXing.Net/releases. Fortunatly zxing had compiled version of .net 2.0. This saved lot of time for me.